Send PrepareOk after state transfer
The simulator found a corner case where we fail to acknowledge a write
and, therefore, not apply the state machine on the primary.

First, the client receives a request, but `Prepare` message never
arrives to one of the replicas:

  Primary <- Request(req=0)
  Backup 1 <- Prepare(op=1)
  Backup 2 <- DROP Prepare(op=1)
  Primary <- PrepareOk(op=1)

As quorum is reached, so primary applies the op to the state machine,
and completes the request.

However, now another request arrives to the client, but this time
`Prepare` message doesn't arrive to the *other* replica. But the replica
that does receive the `Prepare` message is behind in the log, so it
initiates a state transfer:

  Primary <- Request(req=1)
  Backup 1 <- DROP Prepare(op=2)
  Backup 2 <- Prepare(op=2)
  Primary <- GetState(id=2)
  Backup 2 <- NewState(op=2)

However, after the state transfer, we must acknowledge the op; otherwise
primary will just keep waiting forever. Please note that if we did
heartbeats, this issue would be solved once a timer fires, but we should
acknowledge the op immediately anyway.